+++ docs/index.js
@@ -1,7 +1,12 @@
 var React = require('react');
 var {Calendar} = require('../src');
 
+function datePicked(date) {
+  console.log(date);
+}
+
 React.render(
-	<Calendar showDaysOfWeek={true} />,
+	<Calendar showDaysOfWeek={true}
+            onPickDate={datePicked} />,
 	document.getElementById('calendar')
 );