+++ src/Day.js
@@ -10,9 +10,14 @@ var Day = React.createClass({
     };
   },
 
+  _onClick: function() {
+    if (this.props.onClick)
+      this.props.onClick(this.props.day.day);
+  },
+
   render: function() {
     return (
-      <div className={this.props.day.classes}>
+      <div onClick={this._onClick} className={this.props.day.classes}>
         <span className='day-number'>{this.props.day.day.date()}</span>
       </div>
     );