+++ CHANGELOG.md
@@ -1,2 +1,7 @@
+
+## v0.2.0
+* Added week day headers (optional)
+* Added onPickDate callback
+
 ## v0.1.0
 `require('react-calendar-component')` returns an object with all the elements instead of just `<Calendar/>`.